Moved RustDeserializationError to src/deserialize.rs.
