+++ src/deserialize.rs
@@ -7,7 +7,6 @@ use serde::de::{
     VariantAccess, Visitor,
 };
 
-use crate::error::RustDeserializationError;
 use crate::identifier::Ident;
 use crate::term::{MetaValue, RichTerm, Term};
 
@@ -43,6 +42,18 @@ macro_rules! deserialize_number_round {
             }
         }
     };
+}
+
+/// An error occurred during deserialization to Rust.
+#[derive(Debug, PartialEq, Clone)]
+pub enum RustDeserializationError {
+    InvalidType { expected: String, occurred: String },
+    MissingValue,
+    EmptyMetaValue,
+    UnimplementedType { occurred: String },
+    InvalidRecordLength(usize),
+    InvalidArrayLength(usize),
+    Other(String),
 }
 
 impl<'de> serde::Deserializer<'de> for RichTerm {
@@ -543,6 +554,40 @@ impl<'de> EnumAccess<'de> for EnumDeserializer {
             rich_term: self.rich_term,
         };
         seed.deserialize(variant).map(|v| (v, visitor))
+    }
+}
+
+impl std::fmt::Display for RustDeserializationError {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            RustDeserializationError::InvalidType {
+                ref expected,
+                ref occurred,
+            } => write!(f, "invalid type: {occurred}, expected: {expected}"),
+            RustDeserializationError::MissingValue => write!(f, "missing value"),
+            RustDeserializationError::EmptyMetaValue => write!(f, "empty Metavalue"),
+            RustDeserializationError::InvalidRecordLength(len) => {
+                write!(f, "invalid record length, expected {len}")
+            }
+            RustDeserializationError::InvalidArrayLength(len) => {
+                write!(f, "invalid array length, expected {len}")
+            }
+            RustDeserializationError::UnimplementedType { ref occurred } => {
+                write!(f, "unimplemented conversion from type: {occurred}")
+            }
+            RustDeserializationError::Other(ref err) => write!(f, "{err}"),
+        }
+    }
+}
+
+impl std::error::Error for RustDeserializationError {}
+
+impl serde::de::Error for RustDeserializationError {
+    fn custom<T>(msg: T) -> Self
+    where
+        T: std::fmt::Display,
+    {
+        RustDeserializationError::Other(msg.to_string())
     }
 }
